Deepen Keystone kernel and Mason registry
