+++ AGENTS.md
@@ -31,8 +31,11 @@ This repository is the early Keystone UI monorepo. Treat `Keystone` and `Mason` 
 - [docs/adr/](docs/adr/): durable decisions.
 - [docs/agents/](docs/agents/): agent/work tracking conventions.
 - [docs/adr/0003-mason-tanstack-app-layer.md](docs/adr/0003-mason-tanstack-app-layer.md): TanStack app-layer decision for Mason.
+- [docs/prd/keystone-mason-foundation.md](docs/prd/keystone-mason-foundation.md): foundation PRD and product flywheel.
+- [docs/prd/keystone-internals-inspiration-parity.md](docs/prd/keystone-internals-inspiration-parity.md): Keystone kernel parity PRD.
+- [docs/agents/issue-triage-spine.md](docs/agents/issue-triage-spine.md): GitHub issue milestones, labels, and 0.1 active spine.
 - [docs/agents/end-state-primitive-component-inventory.md](docs/agents/end-state-primitive-component-inventory.md): optimal Keystone primitive and Mason component inventory.
-- `.context/attachments/pasted_text_2026-05-03_14-26-48.txt`: source PRD for the current strategic direction.
+- [docs/agents/data-dense-workspace-verticals.md](docs/agents/data-dense-workspace-verticals.md): Mason workspace-pattern issue breakdown.
 
 ## Implementation Guardrails
 
@@ -43,6 +46,7 @@ This repository is the early Keystone UI monorepo. Treat `Keystone` and `Mason` 
 - Overlay work must account for focus management, dismissal, layering, portals, SSR, hydration, and accessibility testing.
 - Mason generated code should be readable source owned by the user project and should not reimplement Keystone behavior.
 - Mason form, table, store, and hotkey work should prefer TanStack libraries instead of custom app-behavior systems.
+- Mason data-dense components should preserve clarity under frequent updates and should treat keyboard navigation/focus behavior as first-class product concerns.
 - Every first-party Mason primitive/component/block registry item should carry `meta.parity` notes. Prefer Base UI first for runtime depth and Kobalte second for Solid-native primitive shape; use the more fitting reference for exceptions such as TanStack app components, Sonner-style toast behavior, Mason utilities, or shadcn-style source registry conventions.
 - When updating the root `README.md`, keep it external-facing and grounded in current repo state. Do not turn it into an internal roadmap; link or defer durable planning detail to `CONTEXT.md`, `CONTEXT-MAP.md`, ADRs, RFCs, and `docs/agents/`.
