+++ packages/keystone/src/tabs/index.tsx
@@ -18,7 +18,7 @@ import {
   renderPolymorphic,
   type PolymorphicProps,
 } from "../utils/index";
-import { useDirection, type Direction as KeystoneDirection } from "../direction/index";
+import { useDirection, type Direction as KeystoneDirection } from "../i18n/direction";
 
 export type TabsActivationMode = "automatic" | "manual";
 export type TabsDirection = KeystoneDirection;
